refactor: implement handlers
Moves the logic to isolated handler functions, cleaning up the main function
