+++ src/core/mod.rs
@@ -1,4 +1,6 @@
 //! The main module with functions to parse, convert, and sort snippets
 pub mod composer;
+pub mod convert_handler;
 pub mod friendly_tekton;
 pub mod snipmate_tekton;
+pub mod sort_handler;
